requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
python-multipart==0.0.*
sqlalchemy==2.0.*
pydantic-settings==2.7.*
openai==1.60.*
pdfplumber==0.11.*
spacy==3.8.*
# English pipeline for PII/entity detection in redactor — not installed with spacy itself:
#   python -m spacy download en_core_web_sm
openpyxl==3.1.*
pandas==2.2.*
